Implement Glide
*Transition image loading from internal loading mechanism to Glide v4.0.0. Glide now performs better than the internal loading mechanism.
